gmail: restructure to move util functions out
